[ALPS04705831] ion: fix code defect
[Detail]
null pointer handling
print with correct argument type

MTK-Commit-Id: 25711e0b40abefbe0660a0b7b08c4e2b8b2ba62e

CR-Id: ALPS04705831
Feature: [Module]ION/M4U

+++ drivers/staging/android/ion/ion_system_heap.c
@@ -93,7 +93,7 @@ static void free_buffer_page(struct ion_system_heap *heap,
 		__free_pages(page, order);
 		if (atomic64_sub_return((1 << order), &page_sz_cnt) < 0) {
 			seq_printf(NULL, "underflow!, total[%lu]free[%lu]\n",
-				   atomic64_read(&page_sz_cnt),
+				   (unsigned long)atomic64_read(&page_sz_cnt),
 				   (unsigned long)(1 << order));
 			atomic64_set(&page_sz_cnt, 0);
 		}
